Initial checkin of Game Manager app
This is a basic implementation of a Game Manager for Android. The functionality is pretty basic - it can list out all of the games on a user's device, sorting by name, install time, or last play time.

Note - in order to support the "recently played" field, the app needs access to the Usage Stats API. This can only be granted by the user, so this requires a dialog at runtime.